services/profile-service/requirements.txt
Django==5.1.0
djangorestframework==3.15.0
drf-spectacular==0.27.2
django-environ==0.11.2
psycopg2-binary==2.9.9
gunicorn==22.0.0
uvicorn==0.30.1
celery==5.4.0
Pillow==10.3.0
django-storages==1.14.2
boto3==1.34.0
django-cors-headers==4.1.0
django-redis==5.4.0